[project @ 71e40e36ebd86fbf541d73ab80b822a22b8c7d0d] [unstable] $mailcap_sanitize.
Original author: Thomas Roessler <roessler@does-not-exist.org>
Date: 1999-02-10 21:42:05+00:00

git-svn-id: http://oss.sygroup.ch/svn/mutt/trunk@814 2bb92213-1a62-4c20-bdc6-606bde718ed5

+++ pager.c
@@ -85,8 +85,8 @@ struct q_class_t
 struct syntax_t
 {
   int color;
-  int first;
-  int last;
+  short first;
+  short last;
 };
 
 struct line_t
